+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 gyselroth Helper Library
 ========================
 
-Version 0.3.00 - Not released yet
----------------------------------
+Version 0.3.0 - Not released yet
+--------------------------------
 * Extract HtmlPurifier filters into separate dependency
 * Remove wrapper methods from HelperString redirecting to HelperDate 
 * Remove wrapper methods from HelperString redirecting to HelperFile